+++ tests/testthat/test-dplyr.R
@@ -259,6 +259,11 @@ test_that("basic test: 'join()'-methods for `zoomed.dm` work", {
     anti_join(tf_2(), tf_1(), by = c("d" = "a"))
   )
 
+  expect_equivalent_tbl(
+    nest_join(zoomed_dm(), tf_1) %>% dm_update_zoomed() %>% tbl_impl("tf_2"),
+    nest_join(tf_2(), tf_1(), by = c("d" = "a"), name = "tf_1")
+  )
+
   # SQLite doesn't implement right join
   skip_if_src("sqlite")
   skip_if_src("maria")
@@ -408,9 +413,13 @@ test_that("basic test: 'join()'-methods for `dm` throws error", {
     "table_not_in_dm"
   )
 
-  skip("No nest_join() for now")
   expect_dm_error(
     nest_join(dm_for_filter()),
+    "only_possible_w_zoom"
+  )
+
+  expect_dm_error(
+    pack_join(dm_for_filter()),
     "only_possible_w_zoom"
   )
 })
@@ -845,6 +854,10 @@ test_that("output for compound keys", {
     # anti_join()
     zoomed_comp_dm %>%
       anti_join(flights) %>%
+      nrow()
+    # nest_join()
+    zoomed_comp_dm %>%
+      nest_join(flights) %>%
       nrow()
   })
 })